requirements.txt
tensorflow-cpu==2.15.0
numpy>=1.23.5,<2.0
pillow>=9.4.0
streamlit>=1.32.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0